Change: Rename Row module to RowDetailsToggle and remove named export
The Row module only contains the RowDetailsToggle component. Therefore
the name of the module is adjusted. Because the module only has a single
export using the default export only is cleaner.
