+++ src/web/entities/EntityNameTableData.jsx
@@ -11,7 +11,7 @@ import Comment from 'web/components/comment/Comment';
 import Layout from 'web/components/layout/Layout';
 import DetailsLink from 'web/components/link/DetailsLink';
 import TableData from 'web/components/table/Data';
-import {RowDetailsToggle} from 'web/entities/Row';
+import RowDetailsToggle from 'web/entities/RowDetailsToggle';
 import ObserverIcon from 'web/entity/icon/ObserverIcon';
 import useUserName from 'web/hooks/useUserName';
 import PropTypes from 'web/utils/PropTypes';
