+++ src/web/pages/dfncert/Row.jsx
@@ -10,7 +10,7 @@ import DateTime from 'web/components/date/DateTime';
 import TableData from 'web/components/table/Data';
 import TableRow from 'web/components/table/Row';
 import EntitiesActions from 'web/entities/Actions';
-import {RowDetailsToggle} from 'web/entities/Row';
+import RowDetailsToggle from 'web/entities/RowDetailsToggle';
 import PropTypes from 'web/utils/PropTypes';
 import {na} from 'web/utils/Render';
