+++ src/web/pages/results/Row.jsx
@@ -22,11 +22,10 @@ import Qod from 'web/components/qod/Qod';
 import TableData from 'web/components/table/Data';
 import TableRow from 'web/components/table/Row';
 import EntitiesActions from 'web/entities/Actions';
-import {RowDetailsToggle} from 'web/entities/Row';
+import RowDetailsToggle from 'web/entities/RowDetailsToggle';
 import useGmp from 'web/hooks/useGmp';
 import ResultDelta from 'web/pages/results/Delta';
 import PropTypes from 'web/utils/PropTypes';
-
 
 const Row = ({
   actionsComponent: ActionsComponent = EntitiesActions,
