+++ src/web/pages/tasks/Row.jsx
@@ -19,7 +19,7 @@ import DetailsLink from 'web/components/link/DetailsLink';
 import Link from 'web/components/link/Link';
 import TableData from 'web/components/table/Data';
 import TableRow from 'web/components/table/Row';
-import {RowDetailsToggle} from 'web/entities/Row';
+import RowDetailsToggle from 'web/entities/RowDetailsToggle';
 import ObserverIcon from 'web/entity/icon/ObserverIcon';
 import useUserName from 'web/hooks/useUserName';
 import Actions from 'web/pages/tasks/Actions';
